refactor: address PR review comments for email forwarding
- 1 export per file: split constants, types, job data into individual files
- Extract shared email parsing utils (extractThreadId, extractParticipants,
  extractAddresses) from IMAP and inbound-email drivers into shared utils
- Reuse existing S3 bucket (STORAGE_S3_NAME) instead of separate
  INBOUND_EMAIL_S3_BUCKET; remove redundant config vars
- Remove attachment handling from inbound parser (not managed)
- Remove unnecessary/obvious comments throughout
- Rename "Connect Account" to "Connect via IMAP/SMTP"
- Remove description from email forwarding card for UI consistency
- Move EMAIL_FORWARDING_MODAL_ID to its own constant file
- Remove duplicate channel check (support multiple forwarding channels)
