+++ packages/twenty-front/src/modules/settings/accounts/constants/EmailForwardingModalId.ts
@@ -1,0 +1,1 @@
+export const EMAIL_FORWARDING_MODAL_ID = 'email-forwarding-address-modal';
