+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-local-part-prefix.constant.ts
@@ -1,0 +1,1 @@
+export const INBOUND_EMAIL_LOCAL_PART_PREFIX = 'ch_';
