+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-local-part-random-bytes.constant.ts
@@ -1,0 +1,1 @@
+export const INBOUND_EMAIL_LOCAL_PART_RANDOM_BYTES = 12;
