+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-s3-prefixes.constant.ts
@@ -1,0 +1,6 @@
+export const INBOUND_EMAIL_S3_PREFIXES = {
+  incoming: 'inbound-email/incoming/',
+  processed: 'inbound-email/processed/',
+  unmatched: 'inbound-email/unmatched/',
+  failed: 'inbound-email/failed/',
+} as const;
