+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/constants/x-twenty-origin-header.constant.ts
@@ -1,0 +1,1 @@
+export const X_TWENTY_ORIGIN_HEADER = 'x-twenty-origin';
