+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/services/inbound-email-parser.service.ts
@@ -1,28 +1,17 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 
-import PostalMime, {
-  type Email as ParsedEmail,
-  type Address,
-} from 'postal-mime';
+import PostalMime, { type Email as ParsedEmail } from 'postal-mime';
 import { MessageDirection } from 'src/modules/messaging/common/enums/message-direction.enum';
-import { MessageParticipantRole } from 'twenty-shared/types';
 
-import { type EmailAddress } from 'src/modules/messaging/message-import-manager/types/email-address';
+import { X_TWENTY_ORIGIN_HEADER } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/constants/x-twenty-origin-header.constant';
+import { type ParsedInboundMessage } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/types/parsed-inbound-message.type';
 import { type MessageWithParticipants } from 'src/modules/messaging/message-import-manager/types/message';
-import { formatAddressObjectAsParticipants } from 'src/modules/messaging/message-import-manager/utils/format-address-object-as-participants.util';
+import { extractParticipantsFromParsedEmail } from 'src/modules/messaging/message-import-manager/utils/extract-participants-from-parsed-email.util';
+import { extractThreadIdFromParsedEmail } from 'src/modules/messaging/message-import-manager/utils/extract-thread-id-from-parsed-email.util';
 import { sanitizeString } from 'src/modules/messaging/message-import-manager/utils/sanitize-string.util';
-import { X_TWENTY_ORIGIN_HEADER } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email.constants';
-
-export type ParsedInboundMessage = {
-  parsed: ParsedEmail;
-  originWorkspaceId: string | null;
-  message: MessageWithParticipants;
-};
 
 @Injectable()
 export class InboundEmailParserService {
-  private readonly logger = new Logger(InboundEmailParserService.name);
-
   async parse(
     rawMessage: Buffer,
     s3Key: string,
@@ -52,88 +41,15 @@ export class InboundEmailParserService {
     s3Key: string,
   ): MessageWithParticipants {
     return {
-      // The S3 key is the natural external id: it uniquely identifies the
-      // object that produced this message and survives retries.
       externalId: `inbound-email:${s3Key}`,
-      messageThreadExternalId: this.extractThreadId(parsed),
+      messageThreadExternalId: extractThreadIdFromParsedEmail(parsed),
       headerMessageId: parsed.messageId?.trim() || `inbound-${s3Key}`,
       subject: sanitizeString(parsed.subject || ''),
       text: sanitizeString(parsed.text || ''),
       receivedAt: parsed.date ? new Date(parsed.date) : new Date(),
-      // Forwarding channels are inbound-only — every message is incoming.
       direction: MessageDirection.INCOMING,
-      attachments: (parsed.attachments || []).map((attachment) => ({
-        filename: attachment.filename || 'unnamed-attachment',
-      })),
-      participants: this.extractParticipants(parsed),
+      attachments: [],
+      participants: extractParticipantsFromParsedEmail(parsed),
     };
-  }
-
-  private extractThreadId(parsed: ParsedEmail): string {
-    const references = parsed.references;
-
-    if (typeof references === 'string' && references.trim()) {
-      const first = references.trim().split(/\s+/)[0];
-
-      if (first) {
-        return first;
-      }
-    }
-
-    if (Array.isArray(references) && references.length > 0) {
-      const first = String(references[0]).trim();
-
-      if (first) {
-        return first;
-      }
-    }
-
-    if (parsed.inReplyTo) {
-      const inReplyTo = String(parsed.inReplyTo).trim();
-
-      if (inReplyTo) {
-        return inReplyTo;
-      }
-    }
-
-    if (parsed.messageId?.trim()) {
-      return parsed.messageId.trim();
-    }
-
-    return `thread-${crypto.randomUUID()}`;
-  }
-
-  private extractParticipants(parsed: ParsedEmail) {
-    const addressFields = [
-      { field: parsed.from, role: MessageParticipantRole.FROM },
-      { field: parsed.to, role: MessageParticipantRole.TO },
-      { field: parsed.cc, role: MessageParticipantRole.CC },
-      { field: parsed.bcc, role: MessageParticipantRole.BCC },
-    ] as const;
-
-    return addressFields.flatMap(({ field, role }) =>
-      formatAddressObjectAsParticipants(this.extractAddresses(field), role),
-    );
-  }
-
-  private extractAddresses(
-    address: Address | Address[] | undefined,
-  ): EmailAddress[] {
-    if (!address) {
-      return [];
-    }
-
-    const addresses = Array.isArray(address) ? address : [address];
-
-    const mailboxes = addresses.flatMap((addr) =>
-      addr.address ? [addr] : (addr.group ?? []),
-    );
-
-    return mailboxes
-      .filter((mailbox) => mailbox.address)
-      .map((mailbox) => ({
-        address: mailbox.address,
-        name: sanitizeString(mailbox.name || ''),
-      }));
   }
 }
