+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/types/parsed-inbound-message.type.ts
@@ -1,0 +1,9 @@
+import { type Email as ParsedEmail } from 'postal-mime';
+
+import { type MessageWithParticipants } from 'src/modules/messaging/message-import-manager/types/message';
+
+export type ParsedInboundMessage = {
+  parsed: ParsedEmail;
+  originWorkspaceId: string | null;
+  message: MessageWithParticipants;
+};
