+++ packages/twenty-server/src/modules/messaging/message-import-manager/jobs/messaging-inbound-email-import-job-data.type.ts
@@ -1,0 +1,3 @@
+export type MessagingInboundEmailImportJobData = {
+  s3Key: string;
+};
